Rename scripts folder to pages
